Moving KeyRingMock class to a single file
